feat: [api-handlers-decomposition-recovery-004] - Extract model route handlers
