+++ docs/internal/processes/development-guide-relevant-files.md
@@ -140,7 +140,8 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/api/handlers_work.go` | Work route handlers | Owns `POST /work`, `PUT /work-requests/{requestId}`, `GET /work`, `GET /work/{id}`, staged submit-work file routes, submit/upsert body decode (`decodeSubmitWorkRequestBody`, `decodeWorkRequestBody`), list/get projection helpers (`tokenToWork`, `workMatchesListFilters`), and `upsertWorkRequestResponse`; merged former `submit_work_items.go` to preserve the `pkg/api` 15-file cap. |
 | `pkg/api/handlers_factory.go` | Factory validation route handlers | Owns `POST /factory/validate` (`ValidateFactory`), `decodeNamedFactoryBody`, and factory-definition validation response mapping via `factoryvalidation.Validate` / `ToValidationTargets`. |
 | `pkg/api/handlers_factory_sessions.go` | Factory session and current-factory route handlers | Owns `/factory-sessions` list/open/close, default and session-scoped current-factory GET, `PUT /factory-sessions/{session_id}/factory` (`SaveCurrentFactoryBySessionId` with `{ mode?, factory }` → `SaveFactoryForSession` only), workstation prompt-template contract/validate routes, `loadCurrentFactory*` / `writeCurrentFactoryError`, and session/factory body decode helpers. |
-| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep model, events/status, and shared projection helpers in `handlers.go` while route-family handlers live in `handlers_work.go`, `handlers_factory.go`, `handlers_factory_sessions.go`, and follow-on `handlers_models.go` / `handlers_events.go` splits; shared JSON/error writers live in `handlers_common.go`. Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
+| `pkg/api/handlers_models.go` | Model route handlers | Owns `GET /models`, `GET /models/{model_name}`, `POST /models/{model_name}/invoke`, `POST /models/{model_name}/pull`, plus `decodeModelInvocationRequestBody` and model-invocation response projection helpers. |
+| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep events/status routes and shared token/status projection helpers in `handlers.go` while route-family handlers live in `handlers_work.go`, `handlers_factory.go`, `handlers_factory_sessions.go`, `handlers_models.go`, and follow-on `handlers_events.go`; shared JSON/error writers live in `handlers_common.go`. Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
 | `pkg/api/server_*_test.go` plus `pkg/api/server_test_helpers_test.go` | API route regression ownership | When API route tests cross the backend size gate, split them into same-package sibling `server_*_test.go` files by boundary concern such as submit, upsert, factory, provider-session, work-query, and dashboard/event coverage, and keep shared request builders or response assertions in `server_test_helpers_test.go` so route behavior stays observable without recreating one giant test owner file. |
 | `api/codegen_config/client.yaml` | Generated Go client config | The checked-in Go client for backend functional tests is generated from this config and must stay in lockstep with `api/openapi.yaml`, `pkg/generatedclient/`, and `make api-smoke`. |
 | `api/openapi-main.yaml` | Authored OpenAPI source of truth | Public API behavior changes should start here, including the advertised local `servers` URL, then refresh the bundled spec and generated Go/UI artifacts together so handwritten route seams and generated parameter binding stay aligned. |
@@ -576,7 +577,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Treat packaged-binary release fixtures under `tests/release/testdata/` as canonical authored factory surfaces. If the public factory contract gains a new required field like top-level `factory.name`, update those release fixtures with the same explicit authored value instead of relying on init-time defaults or test-only backfills.
 - When `/factory` contract shapes change, update `api/openapi-main.yaml`, run `make generate-api`, and then reconcile the handwritten wrappers in `pkg/api/handlers_factory_sessions.go`, `pkg/apisurface/contract.go`, `pkg/service/factory.go`, and `ui/src/api/named-factory/api.ts`; those seams still own request decoding and current-factory adapter behavior after generation.
 - When a new runtime discovery route such as `/models` is added, treat `api/openapi-main.yaml`, generated artifacts, `pkg/apisurface/contract.go`, the owning read-model logic in `pkg/service/`, and the paired CLI transport under `pkg/cli/` as one contract seam; generation adds the route types, but the handwritten handler, runtime adapter, and CLI read path still need explicit updates together.
-- When a `/models/{model_name}/...` mutation route is added, reuse the canonical `WorkContent` plus workstation-binding schemas instead of inventing a parallel payload shape, and keep the generated invocation request or response types aligned across request validation in `pkg/api/handlers.go`, direct worker selection or execution in `pkg/service/`, and the matching `pkg/cli/models` transport path.
+- When a `/models/{model_name}/...` mutation route is added, reuse the canonical `WorkContent` plus workstation-binding schemas instead of inventing a parallel payload shape, and keep the generated invocation request or response types aligned across request validation in `pkg/api/handlers_models.go`, direct worker selection or execution in `pkg/service/`, and the matching `pkg/cli/models` transport path.
 - Keep local model execution behind one `pkg/localmodels` manager composed by service: resolve cache paths through `localmodels.AssetPuller.ResolveModelCache(...)`, key loaded handles by canonical MODEL resource metadata (`model|backend|loadPolicy`), and wrap the shared `workers.Runner` path so direct `/models/.../invocations` and factory-level `MODEL_INVOKE` executions reuse the same local load/invoke seam.
 - When `BuildFactoryService(...)` or runtime replacement in `pkg/service/factory.go` creates local-model managers, copy `modelAssets`, `modelResources`, and `localModels` onto the live `FactoryService` as well as the runtime bundle; direct `/models/.../invocations` consult those service-owned fields outside the factory-runner startup path.
 - When managed local-model pull work changes `pkg/localmodels/assets/`, persist enough cache metadata at pull time and make `ResolveModelCache(...)` prefer disk-backed metadata before remote manifest fetches so already-pulled local models still load and invoke offline.
